Adding bearer token in compass (#16)
<!-- begin-generated-description -->

This pull request introduces a new `bearer_token` parameter to the
`__init__` method of the `compass_sdk/compass.py` file. The
`bearer_token` is an optional string that can be used for authentication
purposes.

The changes made in this PR are as follows:
- A new `bearer_token` parameter is added to the `__init__` method,
allowing for optional authentication using a bearer token.
- The `self.bearer_token` attribute is set to the value of the
`bearer_token` parameter.
- The `_send_request_with_retry` method is modified to include a check
for the `self.bearer_token`. If it exists, the request headers are
updated with the bearer token for authentication.

The PR enhances the authentication process by providing an alternative
to the existing username and password-based authentication. The bearer
token can be used to securely authenticate requests, offering a more
flexible and secure approach.

<!-- end-generated-description -->