Add parameters for the new pptx parser (#35)
<!-- begin-generated-description -->

This PR introduces a new class, `PresentationParsingStrategy`, which
extends the `StrEnum` class. This class provides two parsing strategies
for presentations: `Unstructured` and `ImageToMarkdown`. The
`Unstructured` strategy is set as the default.

The `ParserConfig` class is also updated to include a new parameter,
`presentation_parsing_strategy`, which is set to the `Unstructured`
strategy by default.

## Changes:
- Added the `PresentationParsingStrategy` class.
- Added the `presentation_parsing_strategy` parameter to the
`ParserConfig` class.

<!-- end-generated-description -->