Configurable heavy parsing strategy (#41)
Certain document types (i.e., PDF and Slides) have available two parser
versions:

Quick, or "lighter" parsing: processes the documents very fast, e.g.,
with pypdfium2 in the case of PDF, at the cost of semantic elements of
the document being lost (e.g., tables and images).
"Heavy parsers", which use MiniCPM to extract the document text in
Markdown format.
The default strategy for when to use heavy parsing has been, so far, to
trigger it only when quick parsing fails. We call this approach the
Fallback heavy parsing strategy.

In this PR, we add an alternative strategy to define when to use heavy
parsing:

Complementary: runs heavy parsing after quick parsing, and replaces the
document once it is done. This allows the document to be available in
the index for retrieval in very short time, while a richer version will
be available later.