Add type support to update_group_authorization (#70)
The main contribution of this PR is:

- The API used to return the RetryResult object. Now it returns a strong
type.
- It used to be called `edit_group_authorization`. I renamed to
`update_group_authorization` to match the API name.

While at this, I also changed the `_send_request` API to only accept
objects of type BaseModel. This is to reduce its complexity, and also
the unify the way we are sending and receiving objects.