Retry the file if parser fails (#71)
Fixes 
```
ChunkedEncodingError: ('Connection broken: IncompleteRead(4632016 bytes read, 10217602 more expected)', IncompleteRead(4632016 bytes read, 10217602 more expected))
```

This issue was caused by a parser restarting or dying because of some
reason. The request to parser failed and it was not handled gracefully.
This change will retry the file X times (3 by default) and will fail
gracefully with an error log if not able to parse the file

Adding multiple retries to process_file