Remove use of RetryResult for add_attributes (#76)
Like the rest of the APIs, I changed the `add_attributes` method to
return the result (which in this case is None) in case of success or the
error string in case of failures.

I also renamed RetryResult to `_RetryResult` to clearly indicate that
this is an internal class that shouldn't be exposed externally.