chore(send_request): change retry logs to warn (#94)
# Overview

Retries are natural to happen and shouldn't log an error message but a
warning, and only log an error if all retries failed, i.e. we couldn't
make the request. This way we avoid unnecessary error logs if an API is
temporarily unreliable.

# Context

In
[SEA-1941](https://linear.app/cohereai/issue/SEA-1941/huge-volume-of-compass-related-logs)
one of the complaints that North raised was related to retry logs which,
along along with another issue, resulting in a huge volume of logs and
frequent log rotation. The other issue (dumping asset base64 content in
logs) was the major issue. However, to help them (and clients in
general) in focusing on the important logs, we are making the retry logs
warning, unless all retries fail in which cause we emit an error log.
This way the logs dashboard for errors don't show transient error that
didn't actually cause any issue.