feat(examples): Add "examples" sub-project (#95)
Frequently, we need to write example code for using the SDK or share it
with our clients. To save our time, and ensure we always have
high-quality code to share with customers, I created a sub-folder called
"examples" that make use of the SDK to achieve various things.
Currently, it only contains a simple example, but will add more in
follow up contribution, and more examples are expected to come in the
future.