fix(parsing): [CMPS-28] Surface parsing errors (#102)
Update the `process_file` API to raise an exception in case there is an
issue with parsing. The `process_files` API, which internally makes
parallel calls to `process_file`, was then updated to capture exceptions
raised by `process_file` and return them to the user so the user can
track failed files.

This change also impacts the `process_folder` since it internally calls
the `process_files` API.

This change is not backward compatible, and hence the minor version was
bumped.