fix: Remove required `/api` in url as not every deployment would have it (#108)
## Description
This PR introduces a new parameter, `include_api_in_url`, to the
`CompassRootClient` and `CompassClient` classes. This parameter allows
users to specify whether the `/api` segment should be included in the
base URL. The default value is set to `True` to maintain backward
compatibility.

## Changes
- **Added `include_api_in_url` Parameter**: A new optional parameter,
`include_api_in_url`, has been added to the `__init__` methods of both
`CompassRootClient` and `CompassClient`. This parameter controls whether
the `/api` segment is included in the base URL.
- **Updated Base URL Construction**: The construction of the `base_url`
has been updated to conditionally include `/api` based on the value of
`include_api_in_url`.
- **Updated API Endpoints**: In `CompassClient`, the API endpoints have
been updated to use an `f-string` with the `base_api` variable, which is
determined by the `include_api_in_url` parameter. This ensures that the
`/api` segment is only included when necessary.