chore: unpin tenacity dep version (#109)
<!-- begin-generated-description -->

This PR updates the Poetry version from 1.8.4 to 2.0.1 and removes the
`groups` and `markers` fields from the `poetry.lock` file. It also
updates the `tenacity` dependency in the `pyproject.toml` file to use a
caret range instead of an exact version.

- Updates the Poetry version from 1.8.4 to 2.0.1
- Removes the `groups` field from the `poetry.lock` file for all
packages
- Removes the `markers` field from the `poetry.lock` file for the
`colorama` and `exceptiongroup` packages
- Updates the `tenacity` dependency in the `pyproject.toml` file to use
a caret range instead of an exact version
- Updates the `lock-version` field in the `poetry.lock` file from `2.1`
to `2.0`
- Updates the `content-hash` field in the `poetry.lock` file to reflect
the changes made

<!-- end-generated-description -->