feat: Kick off v2 - httpx, better error handling, and refactoring (#118)
Kicking off v2 for Cohere Compass SDK to address some of the common
issues, namely:

- Poor error reporting (sometimes raising an exception, other times
returning an error response)
- Lack of support for asyncio
- General code quality improvement

This PR addresses the first and the last (not completely yet). A follow
up PR will be made to support async.

This is targeting the `v2` branch. Once the branch is mature, we will
merge it into main to make v2 the default.