fix: Add "path" field to RetrievedChunk (#164)
To stay in line with the API ([private repo
ref](https://github.com/cohere-ai/compass/blob/main/api/src/api/model.py#L74)),
I am adding the `path` attribute to the RetrievedChunk. This should
solve an issue where a call to `get_document` doesn't set the `path`
field in the chunks for the user, even though the API is returning it.